fix(crypto): disallow overwriting identity/address
chore(commons): extract callsite to module
feat(commons): read inline source map
fix(commons): Typeserializer bug checking “in”

+++ shared/crypto/lib/Address.ts
@@ -63,6 +63,9 @@ export default class Address {
 
     const json = TypeSerializer.stringify(this.toJSON(), { format });
     const filepath = path.resolve(basePath, `${filename || this.bech32}.json`);
+    if (fs.existsSync(filepath)) {
+      throw new Error('You attempted to overwrite an existing address!!');
+    }
     await fs.promises.writeFile(filepath, json, { encoding: 'utf8' });
     return filepath;
   }
